+++ src/main/java/uea/atena_api/services/ProvaAlunoService.java
@@ -1,6 +1,7 @@
 package uea.atena_api.services;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
@@ -8,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import uea.atena_api.models.ProvaAluno;
+import uea.atena_api.models.SpecialOperation;
+import uea.atena_api.models.enums.SpecialOperations;
 import uea.atena_api.repositories.ProvaAlunoRepository;
 
 @Service
@@ -28,8 +31,14 @@ public class ProvaAlunoService {
 		return provaAlunoRepository.save(correcao);
 	}
 
-	public void deletar(Long codigo) {
-		provaAlunoRepository.deleteById(codigo);
+	public void deletar(Long codigo, SpecialOperation specialOperation) {
+		if(Objects.isNull(specialOperation.getSpecialOperation())) {
+			throw new RuntimeException("Não é possível apagar Entidade porque possui relações.");			
+		}
+		if(specialOperation.getSpecialOperation() == SpecialOperations.DELETE_RELETED) {
+			provaAlunoRepository.deleteById(codigo);
+		}
+		throw new RuntimeException("Operação especial incorreta.");
 	}
 
 	public ProvaAluno buscarPorId(Long codigo) {
